+++ packages/zowe-explorer/src/configuration/Definitions.ts
@@ -155,5 +155,6 @@ export namespace Definitions {
         SETTINGS_OLD_SETTINGS_MIGRATED = "zowe.settings.oldSettingsMigrated",
         V1_MIGRATION_STATUS = "zowe.v1MigrationStatus",
         DS_SEARCH_OPTIONS = "zowe.dsSearchOptions",
+        DISPLAY_RELEASE_NOTES_VERSION = "zowe.displayReleaseNotes",
     }
 }